requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
sqlalchemy
alembic
psycopg[binary]
pydantic
python-multipart
pytest
httpx

